backend/requirements.txt
anyio==4.11.0
asgiref==3.10.0
billiard==4.2.1
CacheControl==0.14.3
cachetools==6.2.0
celery==5.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.2
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.12.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.9.0
django-filter==24.3
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
firebase_admin==7.1.0
flower==2.0.1
google-api-core==2.26.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
msgpack==1.1.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
phonenumbers==8.13.52
pillow==12.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14